Make proxy logging more customizable.
It is sometimes possible to send enough data (in the HTTP
method, requset path, query string or user-agent) to cause
the log line to be truncated such that important information
is lost.

This change allows you to better configure on a per section
basis how pieces of the log line are truncated.
This allows the operator to keep logs in a consistant format,
preventing the loss of important data.

Change-Id: Id53bdd60c37f5f45570975ebb0d71f272c354b83